+++ frontend/src/components/settings/Permissions.vue
@@ -5,6 +5,7 @@
     <ToggleSwitch class="item" v-model="permissions.download" :name="$t('settings.permissions.download')" />
     <ToggleSwitch class="item" v-model="permissions.modify" :name="$t('settings.permissions.modify')" />
     <ToggleSwitch class="item" v-model="permissions.create" :name="$t('settings.permissions.create')" />
+    <ToggleSwitch class="item" v-model="permissions.archive" :name="$t('settings.permissions.archive')" />
     <ToggleSwitch class="item" v-model="permissions.delete" :name="$t('settings.permissions.delete')" />
     <ToggleSwitch class="item" v-model="permissions.admin" :name="$t('settings.permissions.admin')" />
     <ToggleSwitch class="item" v-model="permissions.share" :name="$t('settings.permissions.share')" />